fix: prompt to create config file in 'gwt config view'
When the config file doesn't exist, 'gwt config view' now prompts
the user to create it interactively, instead of just showing an error.

This fixes issue #29 where the config file initialization was not
working for the config view command, even though the interactive
setup functionality already existed and was used by other commands.

The fix uses Config::interactive_setup() to prompt the user when
the config file is missing, then continues to display the config
file contents after successful setup.